// txmatching/web/frontend/src/app/generated/model/countryGenerated.ts
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CountryGenerated {
    country_code: string;
    donor_count: number;
    recipient_count: number;
}