// txmatching/web/frontend/src/app/generated/model/matchingGenerated.ts
/**
 * API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CountryGenerated } from './countryGenerated';
import { RoundGenerated } from './roundGenerated';


export interface MatchingGenerated {
    count_of_transplants: number;
    countries: Array<CountryGenerated>;
    order_id: number;
    rounds: Array<RoundGenerated>;
    score: number;
}